refactor(data-warehouse): move person-property update consumer into warehouse_sources
